refactor: move web research into shared module
