+++ packages/@n8n/ai-utilities/src/web-research/index.ts
@@ -1,0 +1,2 @@
+export { sanitizeWebContent, wrapUntrustedData } from './sanitize-web-content';
+export type { FetchedPage, WebResearchService, WebSearchResponse, WebSearchResult } from './types';
