+++ packages/@n8n/instance-ai/src/index.ts
@@ -106,7 +106,6 @@ export type {
 	SerializableAgentState,
 	Thread,
 } from '@n8n/agents';
-export { sanitizeWebContent, wrapUntrustedData } from './tools/web-research/sanitize-web-content';
 export type { Logger } from './logger';
 export type { CompactionInput } from './compaction';
 export const generateCompactionSummary: typeof CompactionMod.generateCompactionSummary =
