+++ packages/cli/src/modules/agents/agents.service.ts
@@ -52,8 +52,8 @@ import { ActiveExecutions } from '@/active-executions';
 import { CredentialsService } from '@/credentials/credentials.service';
 import { ConflictError } from '@/errors/response-errors/conflict.error';
 import { NotFoundError } from '@/errors/response-errors/not-found.error';
-import { resolveBuiltinNodeDefinitionDirs } from '@/modules/instance-ai/node-definition-resolver';
-import { WebResearchService } from '@/modules/instance-ai/web-research/web-research.service';
+import { resolveBuiltinNodeDefinitionDirs } from '@/node-catalog/node-definition-resolver';
+import { WebResearchService } from '@/modules/web-research/web-research.service';
 import { EphemeralNodeExecutor } from '@/node-execution';
 import type { PubSubCommandMap } from '@/scaling/pubsub/pubsub.event-map';
 import { Publisher } from '@/scaling/pubsub/publisher.service';
