+++ packages/cli/src/modules/web-research/vendor.d.ts
