+++ packages/cli/src/node-catalog/node-definition-resolver.ts
